Modify the grapher so that: - it knows about exclusive dependencies. - it knows about end of streams
If a stream S1 depends exclusively on a non-finished stream S2, we make
S1 the exclusive descendant of S2, and S2's children become S1.

If S2 was finished however, we act as if the stream is a sibling of S2,
so S1 will be under S2 but in a new branch.